Feat: BaseEnitiy 추가
