+++ src/main/java/com/team3/devinit_back/common/BaseEntity.java
@@ -1,0 +1,30 @@
+package com.team3.devinit_back.common;
+
+import jakarta.persistence.*;
+import lombok.Getter;
+
+import java.time.LocalDateTime;
+
+@MappedSuperclass
+public abstract class BaseEntity {
+
+
+    @Getter
+    @Column(updatable = false)
+    private LocalDateTime createdAt;
+
+    @Getter
+    private LocalDateTime updatedAt;
+
+    @PrePersist
+    public void prePersist() {
+        LocalDateTime now = LocalDateTime.now();
+        createdAt = now;
+        updatedAt = now;
+    }
+
+    @PreUpdate
+    public void preUpdate() {
+        updatedAt = LocalDateTime.now();
+    }
+}
